% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_geo_metadata.R
\docType{data}
\name{full_geo_metadata}
\alias{full_geo_metadata}
\title{Equivalencias de codigos de identificacion geografica}
\format{Un tibble con 526 filas y 8 variables:
\describe{
  \item{codprov, coddepto}{identificadores utilizados por INDRA para escrutinios provisorios de elecciones nacionales}
  \item{VAR_censo}{identificadores de unidades geograficas del INDEC}
  \item{VAR_iso}{identificadores de International Organization for Standardization}
  
  ...
}}
\usage{
full_geo_metadata
}
\description{
Contiene identificadores geograficos para 525 departametnos de Argentina
}
\keyword{datasets}
