% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_variables.R
\docType{package}
\name{geofaceteAR}
\alias{geofaceteAR}
\alias{geofaceteAR-package}
\title{\code{geofaceteAR} package}
\description{
Grillas como mapas para Argentina y sus provincias
See the README on
\href{https://github.com/electorArg/geofaceteAR/blob/master/README.md}{Github}
}
