% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid.R
\name{get_grid}
\alias{get_grid}
\title{Carga grillas de distritos de Argentina}
\usage{
get_grid(distrito = NULL)
}
\arguments{
\item{distrito}{un character con el nombre del distrito que se quiere descargar. Disponibles grillas para Argentina
y para cada provincia. Se pueden chequear los parametros con \code{\link{show_arg_codes}}.}
}
\description{
Funcion que descaga grillas (\emph{facet}) para acomodarlas como si fueran mapas de provincias y deparamentos
de Argentina.
}
\examples{

get_grid("TUCUMAN")

}
