% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grillas_geofacet.R
\docType{data}
\name{grillas_geofacet}
\alias{grillas_geofacet}
\title{Base de datos de grillas}
\format{Una lista que contiene 25 `data.frames` con 5 variables:
\describe{
  \item{code_provincia}{identificación del distrito al que pertenece la grilla}
  \item{col}{la longirud relativa de los distritos de segundo nivel en la grilla asemejando la geografía}
  \item{row}{la longirud relativa de los distritos de segundo nivel en la grilla asemejando la geografía}
  \item{code}{cóigo de identificación de cada uno de los distritos que componen la grilla}
  \item{name}{conombre de cada uno de los distritos que componen la grilla}
  
  ...
}}
\usage{
grillas_geofacet
}
\description{
Grillas de Argentina para usar con \code{facet_geo()}
}
\keyword{datasets}
