% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\docType{data}
\name{names}
\alias{names}
\title{Nombres de distritos con grillas disponibles}
\format{Un data frame con 25 filas y 1 variable:
\describe{
  \item{Distrito}{identificador para descargar la grilla del distrito de interes}
  
  ...
}}
\usage{
names
}
\description{
Un archivo de datos que contiene el `id` para llamar a las grillas de Argetina o alguna de las 24 provincias
}
\keyword{datasets}
