% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_prov.R
\name{recode_prov}
\alias{recode_prov}
\title{Recodifica Codigos de Provincias para grilla de Argentina}
\usage{
recode_prov(data, type = NULL)
}
\arguments{
\item{data}{debe ser un 'data.frame' obtenido con \code{\link{get_grid}}. Las grillas grillas disponibles 
se pueden chequear con \code{\link{show_arg_grids}}.}

\item{type}{cual es el codgo que se quiere definir en la grilla. Las opciones son \code{'indra'}, \code{'indec'} o \code{'iso'}}
}
\description{
Funcion que permite recodificar codigo de provincia para utilizacion de grilla de Argentina y hacer mas sencilla
la vinculacion con distinto tipo de bases de datos con diferentes indicadores. El valor por defecto es \code{codprov}
que corresponde a la codificacion de los escrutinios provisorios. Se puede optar por la nomenclautra de 
indec \code{codprov_censo} o \code{ISO 3166-2}.
}
\details{
Respecto los datos se puede consultar la documentación de 
\href{https://www.iso.org/obp/ui/#iso:code:3166:AR}{\code{ISO 3166-2} - International Organization for Standardization} y del
del \href{https://www.iso.org/obp/ui/#iso:code:3166:AR}{\emph{INDEC}}. \code{codprov} es la codificacion con la que se publciaron las bases de datos 
de INDRA, empresa encargada por muchos años de la tarea y utilizados en la libería \href{https://electorarg.github.io/polAr/}{polAr}.
}
