% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_arg_codes.R
\name{show_arg_codes}
\alias{show_arg_codes}
\title{Diccionario de codigos identificadores de distritos}
\usage{
show_arg_codes(viewer = TRUE, nivel = "provincias")
}
\arguments{
\item{viewer}{Por default es \code{TRUE}. Cuando \code{FALSE} imprime en consola.}

\item{nivel}{Opcion para ver diccionario a nivel de \code{"provincias"} o de \code{"departamentos"}.}
}
\description{
Funcion que devuelve un data.frame con codigos y equivalencias de identificacion de unidades geograficas
}
\details{
La opción \code{viewer = TRUE} muestra un tabulado formateado en el _Viewer_ de _RStudio_.
}
\examples{

show_arg_grids(viewer = FALSE)

}
