% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_arg_grids.R
\name{show_arg_grids}
\alias{show_arg_grids}
\title{Diccionario de grillas (\emph{facets}) disponibles}
\usage{
show_arg_grids(viewer = TRUE)
}
\arguments{
\item{viewer}{Por default es \code{TRUE}. Cuando \code{FALSE} imprime en consola.}
}
\description{
Funcion que devuelve un data.frame con los nombre de distrito disponibles (parametro necesarios para llamar la grilla de interes)
}
